+++ BTDB/ODBLayer/TableInfo.cs
@@ -578,6 +578,7 @@ public class TableInfo
                                 handlerLoad(ref reader, ctx, ref Unsafe.As<Int128, byte>(ref value));
                                 propRefSetter(obj, ref Unsafe.As<Int128, byte>(ref value));
                             });
+                            continue;
                         }
 
                         var stackAllocator = ReflectionMetadata.FindStackAllocatorByType(fieldInfo.Type);